make spam listen to config

+++ automod/spam/max_attachments.go
@@ -2,6 +2,6 @@ package spam
 
 import "github.com/blackmesadev/discordgo"
 
-func ProcessMaxAttachments(m *discordgo.Message, limit int) (bool, int) {
-	return len(m.Attachments) <= limit, len(m.Attachments)
+func ProcessMaxAttachments(m *discordgo.Message, limit int64) (bool, int64) {
+	return int64(len(m.Attachments)) <= limit, int64(len(m.Attachments))
 }
