+++ automod/spam/max_links.go
@@ -4,7 +4,7 @@ import "regexp"
 
 var domainsRegex = regexp.MustCompile(`https?://[^\s]+`)
 
-func ProcessMaxLinks(message string, limit int) (bool, int) {
-	count := len(domainsRegex.FindAllStringIndex(message, -1))
+func ProcessMaxLinks(message string, limit int64) (bool, int64) {
+	count := int64(len(domainsRegex.FindAllStringIndex(message, -1)))
 	return count <= limit, count
 }
