+++ automod/spam/max_newlines.go
@@ -2,8 +2,8 @@ package spam
 
 import "strings"
 
-func ProcessMaxNewlines(message string, limit int) (bool, int) {
-	count := strings.Count(message, "\n")
+func ProcessMaxNewlines(message string, limit int64) (bool, int64) {
+	count := int64(strings.Count(message, "\n"))
 
 	if count > limit {
 		return false, count
